+++ composeApp/src/commonMain/kotlin/world/hachimi/app/ui/contributor/ReviewListScreen.kt
@@ -2,7 +2,19 @@ package world.hachimi.app.ui.contributor
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.asPaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.navigationBarsPadding
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -73,8 +85,8 @@ private fun NotContributor() {
 private fun Content(vm: ReviewViewModel) {
     Column(
         Modifier.fillMaxSize()
-            .windowInsetsPadding(WindowInsets.navigationBars)
-            .windowInsetsPadding(LocalContentInsets.current)
+            .navigationBarsPadding()
+            .padding(LocalContentInsets.current.asPaddingValues())
             .padding(vertical = 24.dp),
         Arrangement.spacedBy(16.dp)
     ) {