+++ composeApp/src/commonMain/kotlin/world/hachimi/app/ui/home/HomeMainScreen.kt
@@ -1,8 +1,28 @@
 package world.hachimi.app.ui.home
 
-import androidx.compose.animation.*
+import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.core.tween
-import androidx.compose.foundation.layout.*
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
+import androidx.compose.animation.scaleIn
+import androidx.compose.animation.slideInVertically
+import androidx.compose.animation.togetherWith
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.asPaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.navigationBarsPadding
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.widthIn
+import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyHorizontalGrid
 import androidx.compose.foundation.lazy.grid.items
@@ -22,7 +42,11 @@ import androidx.compose.ui.unit.dp
 import org.koin.compose.koinInject
 import org.koin.compose.viewmodel.koinViewModel
 import world.hachimi.app.api.module.SongModule
-import world.hachimi.app.model.*
+import world.hachimi.app.model.GlobalStore
+import world.hachimi.app.model.HomeViewModel
+import world.hachimi.app.model.InitializeStatus
+import world.hachimi.app.model.fromPublicDetail
+import world.hachimi.app.model.fromSearchSongItem
 import world.hachimi.app.nav.Route
 import world.hachimi.app.ui.LocalContentInsets
 import world.hachimi.app.ui.LocalWindowSize
@@ -52,9 +76,10 @@ fun HomeMainScreen(
         screenWidth = LocalWindowSize.current.width
     ) {
         Column(
-            Modifier.fillMaxSize().verticalScroll(rememberScrollState())
-                .windowInsetsPadding(WindowInsets.navigationBars)
-                .windowInsetsPadding(LocalContentInsets.current)
+            Modifier.fillMaxSize()
+                .verticalScroll(rememberScrollState())
+                .navigationBarsPadding()
+                .padding(LocalContentInsets.current.asPaddingValues())
                 .padding(vertical = 24.dp),
             verticalArrangement = Arrangement.spacedBy(AdaptiveListSpacing)
         ) {