+++ composeApp/src/commonMain/kotlin/world/hachimi/app/ui/root/RootScreen.kt
@@ -2,7 +2,17 @@ package world.hachimi.app.ui.root
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.DrawerState
 import androidx.compose.material3.DrawerValue
@@ -131,8 +141,9 @@ private fun CompactScreen(
             ) {
                 Column(
                     Modifier.padding(top = currentSafeAreaInsets().top)
-                        .consumeWindowInsets(WindowInsets.statusBars)
-                        .navigationBarsPadding()
+                        .padding(bottom = currentSafeAreaInsets().bottom)
+////                        .consumeWindowInsets(WindowInsets.statusBars)
+//                        .navigationBarsPadding()
                 ) {
                     Logo(Modifier.padding(top = 16.dp, start = 16.dp, end = 16.dp))
                     Box(Modifier.padding(12.dp)) {