+++ composeApp/src/commonMain/kotlin/world/hachimi/app/ui/search/SearchScreen.kt
@@ -6,12 +6,12 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.asPaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.windowInsetsBottomHeight
 import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.GridItemSpan
@@ -137,7 +137,7 @@ private fun Content(vm: SearchViewModel, global: GlobalStore) {
         }
 
         item(span = { GridItemSpan(maxLineSpan) }) {
-            Spacer(Modifier.navigationBarsPadding().windowInsetsBottomHeight(LocalContentInsets.current))
+            Spacer(Modifier.navigationBarsPadding().padding(LocalContentInsets.current.asPaddingValues()))
         }
     }
 }